Move bindings for TableLoader and SqlScriptExecutorProvider to the guice support module.
