+++ morf-guicesupport/src/main/java/org/alfasoftware/morf/guicesupport/MorfModule.java
@@ -14,11 +14,20 @@
  */
 package org.alfasoftware.morf.guicesupport;
 
+import javax.sql.DataSource;
+
+import org.alfasoftware.morf.dataset.TableLoader;
+import org.alfasoftware.morf.dataset.TableLoaderBuilder;
+import org.alfasoftware.morf.jdbc.SqlDialect;
+import org.alfasoftware.morf.jdbc.SqlScriptExecutorProvider;
 import org.alfasoftware.morf.upgrade.TableContribution;
 import org.alfasoftware.morf.upgrade.additions.UpgradeScriptAddition;
 import org.alfasoftware.morf.upgrade.db.DatabaseUpgradeTableContribution;
 
+import com.google.common.base.Preconditions;
 import com.google.inject.AbstractModule;
+import com.google.inject.Inject;
+import com.google.inject.Provides;
 import com.google.inject.multibindings.Multibinder;
 
 /**
@@ -37,5 +46,43 @@ public class MorfModule extends AbstractModule {
     Multibinder<TableContribution> tableMultibinder = Multibinder.newSetBinder(binder(), TableContribution.class);
     tableMultibinder.addBinding().to(DatabaseUpgradeTableContribution.class);
   }
-}
 
+
+  /**
+   * Provides a {@link SqlScriptExecutorProvider}.
+   *
+   * @param param The optional bindings.
+   * @return The {@link SqlScriptExecutorProvider}.
+   */
+  @Provides
+  public SqlScriptExecutorProvider sqlScriptExecutorProvider(SqlScriptExecutorProviderParam param) {
+    Preconditions.checkNotNull(param.sqlDialect, "Dialect is null");
+    Preconditions.checkNotNull(param.dataSource, "Data source is null");
+    return new SqlScriptExecutorProvider(param.dataSource, param.sqlDialect);
+  }
+
+  private static final class SqlScriptExecutorProviderParam {
+    @Inject(optional = true) SqlDialect sqlDialect;
+    @Inject(optional = true) DataSource dataSource;
+  }
+
+
+  /**
+   * Provides a {@link TableLoaderBuilder}.
+   *
+   * @param param The optional bindings.
+   * @param sqlScriptExecutorProvider The SQL script executor.
+   * @return The {@link TableLoaderBuilder}.
+   */
+  @Provides()
+  public TableLoaderBuilder tableLoaderBuilder(TableLoaderBuilderParam param, SqlScriptExecutorProvider sqlScriptExecutorProvider) {
+    Preconditions.checkNotNull(param.sqlDialect, "Dialect is null");
+    return TableLoader.builder()
+        .withDialect(param.sqlDialect)
+        .withSqlScriptExecutor(sqlScriptExecutorProvider.get());
+  }
+
+  private static final class TableLoaderBuilderParam {
+    @Inject(optional = true) SqlDialect sqlDialect;
+  }
+}
